Importing changes

+++ archivelib-sys2/c-lib/src/r_expand/fn258.cpp
@@ -3,14 +3,15 @@
 void RExpand::fn258(int32_t _259, uint8_t *_260, int32_t _261, uint16_t *_262,
                     uint16_t _263) {
   uint16_t _277[17], _287[17], _288[18], *_204;
-  uint run_start226, _289, _209, _290, _291, _292, _293, _283;
+  uint32_t run_start226, _289, _209, _290, _291, _292, _293, _283;
   for (run_start226 = 1; run_start226 <= 16; run_start226++)
     _277[run_start226] = 0;
-  for (run_start226 = 0; (int32_t)run_start226 < _259; run_start226++)
+  for (run_start226 = 0; run_start226 < _259; run_start226++)
     _277[_260[run_start226]]++;
   _288[1] = 0;
   for (run_start226 = 1; run_start226 <= 16; run_start226++)
-    _288[run_start226 + 1] = (uint16_t)(_288[run_start226] + (_277[run_start226] << (16 - run_start226)));
+    _288[run_start226 + 1] =
+        _288[run_start226] + (_277[run_start226] << (16 - run_start226));
   if (_288[17] != (uint16_t)(1U << 16)) {
     mStatus.SetError(AL_INTERNAL_ERROR, INTERNAL_ERROR_1_MSG);
     data->error_counter243 = 10;
@@ -33,11 +34,11 @@ void RExpand::fn258(int32_t _259, uint8_t *_260, int32_t _261, uint16_t *_262,
   }
   _292 = _259;
   _283 = 1U << (15 - _261);
-  for (_290 = 0; (int32_t)_290 < _259; _290++) {
+  for (_290 = 0; _290 < _259; _290++) {
     if ((_209 = _260[_290]) == 0)
       continue;
     _293 = _288[_209] + _287[_209];
-    if ((int32_t)_209 <= _261) {
+    if (_209 <= _261) {
       if (_293 > _263) {
         mStatus.SetError(AL_INTERNAL_ERROR, INTERNAL_ERROR_2_MSG);
         data->error_counter243 = 10;
